requirements.txt
chromadb==0.6.3
Flask==3.1.0
PyYAML==6.0.2
Requests==2.32.3
sentence_transformers==3.4.1
Gunicorn==23.0.0